requirements.txt
uvicorn[standard]
fastapi
jinja2
python-jose
sqlalchemy
pydantic-settings>=2.0
pydantic>=2.0
psycopg2-binary
passlib
python-multipart
bcrypt
aiofiles